requirements.txt
pytest>=7.0.1
allure-pytest>=2.9.45
